Expand local positioning replay history

+++ core/middleware/incoming_payloads/service/position_correction_payload.hpp
@@ -8,7 +8,7 @@ namespace middleware_incoming_payloads
 {
   struct position_correction_payload_data
   {
-    std::uint8_t pose_id = 0U;
+    std::uint16_t pose_id = 0U;
     std::uint8_t branch_id = 0U;
   };
